requirements.txt
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
build==1.0.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
coreapi==2.3.3
coreschema==0.0.4
debugpy==1.6.7.post1
decorator==5.1.1
distlib==0.3.7
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
executing==1.2.0
filelock==3.12.3
idna==3.4
ipykernel==6.25.1
ipython==8.14.0
itypes==1.2.0
jedi==0.19.0
Jinja2==3.1.2
jupyter_client==8.3.0
jupyter_core==5.3.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.7
numpy==1.25.2
packaging==23.1
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
pip-tools==7.3.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.16.1
pyproject_hooks==1.0.0
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.1.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tornado==6.3.3
traitlets==5.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
virtualenv==20.24.4
wcwidth==0.2.6
